more streamlined code

+++ vis/utils/io.py
@@ -13,8 +13,9 @@
 
 """
 
-import numpy as np
+import os
 import ast
+import numpy as np
 
 __all__ = ["loadtxt", "savetxt"]
 
@@ -66,13 +67,17 @@ def loadtxt(fname, dtype=float, delimiter=None):
     X : ndarray
         Data read from the text file.
     """
-    fp = open(fname, 'r')
-    X = []
-    for l in fp:
-        a = [cast(v.strip(), dtype) for v in l.strip().split(delimiter)]
-        X.append(np.array(a))
-    fp.close()
-    return np.array(X, dtype=object)
+     
+    if os.path.exists(fname):
+        fp = open(fname, 'r')
+        X = []
+        for l in fp:
+            a = [cast(v.strip(), dtype) for v in l.strip().split(delimiter)]
+            X.append(np.array(a))
+        fp.close()
+        return np.array(X, dtype=object)
+    else:
+        raise OSError("File {0:s} not found.".format(fname))
 
 def savetxt(fname, X, fmt='{:.18e}', delimiter=' ', newline='\n'):
     r"""Save an array to a file, very similar to `npyio.savetxt()`.
@@ -94,8 +99,11 @@ def savetxt(fname, X, fmt='{:.18e}', delimiter=' ', newline='\n'):
     newline : str, optional
         String or character separating lines. Default `\n` when optional.
     """
-    fp = open(fname, 'w+')
-    for i in range(X.shape[0]-1):
-        fp.write(delimiter.join([fmt.format(v) for v in X[i]]) + newline)
-    fp.write(delimiter.join([fmt.format(v) for v in X[X.shape[0]-1]]))
-    fp.close()
+    if os.path.exists(fname):
+        fp = open(fname, 'w+')
+        for i in range(X.shape[0]-1):
+            fp.write(delimiter.join([fmt.format(v) for v in X[i]]) + newline)
+        fp.write(delimiter.join([fmt.format(v) for v in X[X.shape[0]-1]]))
+        fp.close()
+    else:
+        raise OSError("File {0:s} not found.".format(fname))
